requirements/base.txt
Django==4.2
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2-binary==2.9.6
django-jazzmin==2.6.0
python-dotenv==1.0.0
django-recaptcha==3.0.0
drf-yasg==1.21.5
django-modeltranslation==0.18.9
